+++ App/carpi_windows_app/lib/main.dart
@@ -1,11 +1,17 @@
 // ignore_for_file: prefer_const_constructors, prefer_const_literals_to_create_immutables
 
 // import 'package:english_words/english_words.dart';
-import 'package:flutter/material.dart';
 // import 'package:provider/provider.dart';
+
+import 'package:flutter/material.dart';
 import 'package:curved_labeled_navigation_bar/curved_navigation_bar.dart';
 import 'package:curved_labeled_navigation_bar/curved_navigation_bar_item.dart';
 import 'package:google_fonts/google_fonts.dart';
+
+// import the self-made classes
+import 'exercise_page.dart';
+import 'login_page.dart';
+import 'dashboard_page.dart';
 
 void main() {
   runApp(MyApp());
@@ -210,137 +216,6 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 }
 
-class StartupPage extends StatelessWidget {
-  const StartupPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    var theme = Theme.of(context);
-
-    return Container(
-      color: theme.colorScheme.background,
-      child: SafeArea (
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Material (
-              color: theme.colorScheme.background,
-              child: Column (
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    
-                    Text(
-                      "Carpi",
-                      style: TextStyle(color: theme.colorScheme.primary, fontSize: 30, fontWeight: FontWeight.w600),
-                    ),
-                    
-                    Text(
-                      "Rehabilitation",
-                      style: TextStyle(color: theme.colorScheme.primary, fontSize: 30, fontWeight: FontWeight.w100),
-                    ),
-
-                    SizedBox(
-                      width: 300.0,
-                      child: Image.asset('assets/logo.png')
-                      ), //   <--- image
-                  ]
-                ),
-            ),
-            // email field
-            Material (
-              color: theme.colorScheme.background,
-              child : Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: SizedBox(
-                  width: 300.0,
-                child: TextField(
-                  textAlign: TextAlign.center,
-                  keyboardType: TextInputType.text,
-                  decoration: InputDecoration(
-                    hintText: 'Email',
-                      hintStyle: TextStyle(fontSize: 16, fontWeight: FontWeight.w200),
-                    border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(20),
-                        borderSide: BorderSide(
-                            width: 0, 
-                            style: BorderStyle.none,
-                        ),
-                    ),
-                    filled: true,
-                    contentPadding: EdgeInsets.all(16),
-                  ),
-                ),
-                ),
-              ),
-            ),
-
-            // password field
-            Material (
-              color: theme.colorScheme.background,
-              child : Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: SizedBox(
-                  width: 300.0,
-                  child: TextField(
-                    textAlign: TextAlign.center,
-                    keyboardType: TextInputType.text,
-                    obscureText: true,
-                
-                    decoration: InputDecoration(
-                      hintText: 'Password',
-                      hintStyle: TextStyle(fontSize: 16, fontWeight: FontWeight.w200),
-                      border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(20),
-                          borderSide: BorderSide(
-                              width: 0, 
-                              style: BorderStyle.none,
-                          ),
-                      ),
-                      filled: true,
-                      contentPadding: EdgeInsets.all(16),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-
-            // login button
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Center(
-                child: ElevatedButton(
-                  child: const Text("Log in", style:TextStyle(fontWeight: FontWeight.w100)),
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(builder: (context) => MyHomePage()),
-                    );              
-                  },
-                ),
-              ),
-            ),
-
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Center(
-                child: ElevatedButton(
-                  child: const Text("Get started", style:TextStyle(fontWeight: FontWeight.w100)),
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(builder: (context) => MyHomePage()),
-                    );              
-                  },
-                ),
-              ),
-            ),
-        ],
-        ),
-      ),
-    );
-  }
-
-}
 
 // class GeneratorPage extends StatelessWidget {
 
@@ -442,208 +317,6 @@ class StartupPage extends StatelessWidget {
 //   }
 // }
 
-class PrescribedExercises extends StatelessWidget {
-
-    @override
-  Widget build(BuildContext context) {
-    // var theme = Theme.of(context);
-    double width = MediaQuery.of(context).size.width;
-    double height = MediaQuery.of(context).size.height;
-    var padding = MediaQuery.of(context).padding;
-    double newheight = height - padding.top - padding.bottom;
-
-    return SafeArea (
-      child: SizedBox(
-        width: width,
-        height: newheight,
-        child: SingleChildScrollView(
-              child: Column (
-                children: [
-                  
-                  Padding(
-                    padding: const EdgeInsets.all(20.0),
-                    child: SizedBox(
-                      width: double.infinity,
-                      child: RichText(
-                        text: TextSpan(
-                          children: <InlineSpan>[
-                            TextSpan(
-                              text: 'Your',
-                              style: TextStyle(
-                                fontWeight: FontWeight.w100,
-                                fontSize: 30
-                              )
-                            ),
-                            TextSpan(
-                              text: ' Exercises',
-                              style: TextStyle(
-                                fontWeight: FontWeight.w600,
-                                fontSize: 30
-                              )
-                            ),
-                          ]
-                        )
-                      ),
-                    ),
-                  ),
-        
-                  Padding( // insert an element
-                    padding: const EdgeInsets.all(10.0),
-                    child: Card (
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(15), // Adjust the radius as needed
-                      ),
-                      clipBehavior: Clip.antiAlias,
-                      elevation: 5.0,
-                      child: Image.network(
-                        'https://carpirehab.files.wordpress.com/2023/12/draft-1.webp',
-                        width: width,
-                        // height: 250,
-                        fit: BoxFit.contain,
-                      ),
-                    ),
-                  ),
-        
-                  Padding( // insert an element
-                    padding: const EdgeInsets.all(10.0),
-                    child: Card (
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(15), // Adjust the radius as needed
-                      ),
-                      clipBehavior: Clip.antiAlias,
-                      elevation: 5.0,
-                      child: Image.network(
-                        'https://carpirehab.files.wordpress.com/2024/02/1000019337.jpg',
-                        width: width,
-                        // height: 250,
-                        fit: BoxFit.contain,
-                      ),
-                    ),
-                  ),
-        
-                  Padding( // insert an element
-                    padding: const EdgeInsets.all(10.0),
-                    child: Card (
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(15), // Adjust the radius as needed
-                      ),
-                      clipBehavior: Clip.antiAlias,
-                      elevation: 5.0,
-                      child: Image.network(
-                        'https://carpirehab.files.wordpress.com/2024/01/image.png',
-                        width: width,
-                        // height: 250,
-                        fit: BoxFit.contain,
-                      ),
-                    ),
-                  ),
-        
-        
-              ],
-              ),
-            ),
-      ),
-    );
-  }
-}   
-
-class Dashboard extends StatelessWidget {
-
-    @override
-  Widget build(BuildContext context) {
-    // var theme = Theme.of(context);
-    double width = MediaQuery.of(context).size.width;
-    double height = MediaQuery.of(context).size.height;
-    var padding = MediaQuery.of(context).padding;
-    double newheight = height - padding.top - padding.bottom;
-
-    return SafeArea (
-      child: SizedBox(
-        width: width,
-        height: newheight,
-        child: SingleChildScrollView(
-              child: Column (
-                children: [
-                  
-                  Padding(
-                    padding: const EdgeInsets.all(20.0),
-                    child: SizedBox(
-                      width: double.infinity,
-                      child: RichText(
-                        text: TextSpan(
-                          children: <InlineSpan>[
-                            TextSpan(
-                              text: 'Dashboard',
-                              style: TextStyle(
-                                fontWeight: FontWeight.w500,
-                                fontSize: 30
-                              )
-                            ),
-                          ]
-                        )
-                      ),
-                    ),
-                  ),
-        
-                  Padding( // insert an element
-                    padding: const EdgeInsets.all(10.0),
-                    child: Card (
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(15), // Adjust the radius as needed
-                      ),
-                      clipBehavior: Clip.antiAlias,
-                      elevation: 5.0,
-                      child: Image.network(
-                        'https://carpirehab.files.wordpress.com/2023/12/pexels-photo-8386755.jpeg',
-                        width: width,
-                        // height: 250,
-                        fit: BoxFit.contain,
-                      ),
-                    ),
-                  ),
-        
-                  Padding( // insert an element
-                    padding: const EdgeInsets.all(10.0),
-                    child: Card (
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(15), // Adjust the radius as needed
-                      ),
-                      clipBehavior: Clip.antiAlias,
-                      elevation: 5.0,
-                      child: Image.network(
-                        'https://carpirehab.files.wordpress.com/2023/12/pexels-photo-3059748.jpeg',
-                        width: width,
-                        // height: 250,
-                        fit: BoxFit.contain,
-                      ),
-                    ),
-                  ),
-        
-                  Padding( // insert an element
-                    padding: const EdgeInsets.all(10.0),
-                    child: Card (
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(15), // Adjust the radius as needed
-                      ),
-                      clipBehavior: Clip.antiAlias,
-                      elevation: 5.0,
-                      child: Image.network(
-                        'https://carpirehab.files.wordpress.com/2023/12/handstuff.webp?w=2000&h=',
-                        width: width,
-                        // height: 250,
-                        fit: BoxFit.contain,
-                      ),
-                    ),
-                  ),
-        
-        
-              ],
-              ),
-            ),
-      ),
-    );
-  }
-}   
 
 class Progress extends StatelessWidget {
